fix: move alerts to root, fix css
