+++ frontend/components/MainPage/index.tsx
@@ -11,16 +11,14 @@ import { useServices } from '@/hooks/useServices';
 import { useStakingProgram } from '@/hooks/useStakingProgram';
 
 import { MainHeader } from './header';
-import { AddBackupWalletAlert } from './sections/AddBackupWalletAlert';
 import { AddFundsSection } from './sections/AddFundsSection';
+import { AlertSections } from './sections/AlertSections';
 import { GasBalanceSection } from './sections/GasBalanceSection';
 import { KeepAgentRunningSection } from './sections/KeepAgentRunningSection';
 import { MainNeedsFunds } from './sections/NeedsFundsSection';
-import { NewStakingProgramAlertSection } from './sections/NewStakingProgramAlertSection';
 import { MainOlasBalance } from './sections/OlasBalanceSection';
 import { MainRewards } from './sections/RewardsSection';
 import { StakingContractUpdate } from './sections/StakingContractUpdate';
-import { UpdateAlert } from './sections/UpdateAlert';
 
 export const Main = () => {
   const { goto } = usePageState();
@@ -63,11 +61,7 @@ export const Main = () => {
       style={{ borderTopColor: 'transparent' }}
     >
       <Flex vertical>
-        <UpdateAlert />
-        {!backupSafeAddress && <AddBackupWalletAlert />}
-        {currentStakingProgram === StakingProgramId.Alpha && (
-          <NewStakingProgramAlertSection />
-        )}
+        <AlertSections />
         <MainOlasBalance isBorderTopVisible={!hideMainOlasBalanceTopBorder} />
         <MainRewards />
         <KeepAgentRunningSection />
