+++ frontend/styles/globals.scss
@@ -118,7 +118,7 @@ textarea,
     border-left: 0;
     border-right: 0;
     padding: 12px 24px;
-    margin: -24px;
+    margin: 0 -24px;
   }
 }
